spit druid-controller and create fancy mod controller structure
* reflect different "phases" of reconcile function
* clean up imports and error enum variants
