+++ rust/operator-binary/src/parts/mod.rs
@@ -1,0 +1,6 @@
+//! Various parts used in the controller codebase
+
+pub mod config;
+pub mod discovery;
+pub mod extensions;
+pub mod internal_secret;
